+++ info/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Check for COM registry permissions.
 - Helper function for resolving module paths.
 - Check for COM module file permissions.
+- Check for COM ghost DLLs.
 
 ## 2024-06-22